All the Extensions classes are moved to the Extensions namespace.
