+++ src/Cli/Microsoft.DotNet.Cli.Utils/GracefulException.cs
@@ -1,6 +1,8 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Microsoft.DotNet.Cli.Utils.Extensions;
+
 namespace Microsoft.DotNet.Cli.Utils
 {
     public class GracefulException : Exception
