+++ src/Cli/Microsoft.DotNet.Configurer/CliFolderPathCalculator.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 using NuGet.Common;
 
 namespace Microsoft.DotNet.Configurer
