+++ src/Cli/Microsoft.TemplateEngine.Cli/TemplateInvoker.cs
@@ -3,6 +3,7 @@
 
 using System.Text.RegularExpressions;
 using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.TemplateEngine.Abstractions.TemplatePackage;
 using Microsoft.TemplateEngine.Cli.Commands;
