+++ src/Cli/dotnet/ParseResultExtensions.cs
@@ -7,6 +7,7 @@ using System.Diagnostics;
 using System.Reflection;
 using System.Text.RegularExpressions;
 using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 using static Microsoft.DotNet.Cli.Parser;
 using CommandResult = System.CommandLine.Parsing.CommandResult;
