+++ src/Cli/dotnet/ShellShim/WindowsEnvironmentPath.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 using Microsoft.DotNet.Tools;
 
 namespace Microsoft.DotNet.ShellShim
