+++ src/Cli/dotnet/commands/dotnet-msbuild/MSBuildForwardingApp.cs
@@ -6,6 +6,7 @@ using System.Reflection;
 using Microsoft.DotNet.Cli;
 using Microsoft.DotNet.Cli.Telemetry;
 using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 
 namespace Microsoft.DotNet.Tools.MSBuild
 {
