+++ src/Cli/dotnet/commands/dotnet-run/Program.cs
@@ -5,6 +5,7 @@ using System.Collections.Immutable;
 using System.CommandLine;
 using Microsoft.DotNet.Cli;
 using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 using Parser = Microsoft.DotNet.Cli.Parser;
 
 namespace Microsoft.DotNet.Tools.Run
