+++ src/Cli/dotnet/commands/dotnet-run/RunCommand.cs
@@ -9,6 +9,7 @@ using Microsoft.Build.Framework;
 using Microsoft.Build.Logging;
 using Microsoft.DotNet.Cli;
 using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 using Microsoft.DotNet.CommandFactory;
 using Microsoft.DotNet.Tools.Run.LaunchSettings;
