+++ src/Cli/dotnet/commands/dotnet-tool/search/SearchResultPrinter.cs
@@ -3,6 +3,7 @@
 
 using Microsoft.DotNet.Cli;
 using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 
 namespace Microsoft.DotNet.Tools.Tool.Search
 {
