+++ src/Cli/dotnet/commands/dotnet-tool/uninstall/ToolUninstallGlobalOrToolPathCommand.cs
@@ -5,6 +5,7 @@ using System.CommandLine;
 using System.Transactions;
 using Microsoft.DotNet.Cli;
 using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 using Microsoft.DotNet.ShellShim;
 using Microsoft.DotNet.ToolPackage;
 using Microsoft.DotNet.Tools.Tool.Common;
