+++ test/Microsoft.DotNet.ShellShim.Tests/ShellShimRepositoryTests.cs
@@ -8,6 +8,7 @@ using System.Runtime.CompilerServices;
 using System.Transactions;
 using Microsoft.DotNet.Cli.NuGetPackageDownloader;
 using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 using Microsoft.DotNet.ToolPackage;
 using Microsoft.DotNet.Tools;
 using Microsoft.DotNet.Tools.Tests.ComponentMocks;
