+++ test/dotnet.Tests/CommandTests/BuildServerShutdownCommandTests.cs
@@ -5,7 +5,7 @@
 
 using System.CommandLine;
 using Microsoft.DotNet.BuildServer;
-using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.Cli.Utils.Extensions;
 using Microsoft.DotNet.Tools.Test.Utilities;
 using Microsoft.Extensions.EnvironmentAbstractions;
 using Moq;
